% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_entropy.R
\name{conditional_entropy}
\alias{conditional_entropy}
\title{Calculate Conditional Entropy}
\usage{
conditional_entropy(joint_matrix)
}
\arguments{
\item{joint_matrix}{Matrix of joint probabilities. Rows = X outcomes, Cols = Y outcomes.}
}
\value{
Numeric value. Conditional entropy H(Y|X).
}
\description{
Calculate Conditional Entropy
}
\details{
H(Y|X) = H(X,Y) - H(X)
}
