% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_entropy.R
\name{cross_entropy}
\alias{cross_entropy}
\title{Calculate Cross-Entropy}
\usage{
cross_entropy(p, q)
}
\arguments{
\item{p}{Numeric vector. True probability distribution.}

\item{q}{Numeric vector. Estimated probability distribution. Same length as p.}
}
\value{
Numeric value. Cross-entropy H(P, Q).
}
\description{
Calculate Cross-Entropy
}
\details{
Cross-entropy: \eqn{H(P,Q) = -\sum p_i \log q_i}.
}
