% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_entropy.R
\name{joint_entropy}
\alias{joint_entropy}
\title{Calculate Joint Entropy}
\usage{
joint_entropy(joint_matrix)
}
\arguments{
\item{joint_matrix}{Matrix of joint probabilities. Rows = X outcomes, Cols = Y outcomes.}
}
\value{
Numeric value. Joint entropy H(X,Y).
}
\description{
Calculate Joint Entropy
}
