% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_divergence.R
\name{js_divergence}
\alias{js_divergence}
\title{Calculate Jensen-Shannon Divergence}
\usage{
js_divergence(p, q)
}
\arguments{
\item{p}{Numeric vector. First probability distribution.}

\item{q}{Numeric vector. Second probability distribution. Same length as p.}
}
\value{
Numeric value. Jensen-Shannon divergence.
}
\description{
Calculate Jensen-Shannon Divergence
}
\details{
\deqn{JSD(P, Q) = \frac{1}{2} D_{KL}(P || M) + \frac{1}{2} D_{KL}(Q || M)}
where \eqn{M = (P+Q)/2}.
}
