% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_divergence.R
\name{kl_divergence}
\alias{kl_divergence}
\title{Calculate KL Divergence}
\usage{
kl_divergence(p, q)
}
\arguments{
\item{p}{Numeric vector. True probability distribution.}

\item{q}{Numeric vector. Estimated probability distribution. Same length as p.}
}
\value{
Numeric value. KL divergence D_KL(P || Q).
}
\description{
Calculate KL Divergence
}
\details{
\deqn{D_{KL}(P||Q) = \sum p_i \log (p_i / q_i)}
}
