% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_entropy.R
\name{normalized_entropy}
\alias{normalized_entropy}
\title{Calculate Normalized Entropy}
\usage{
normalized_entropy(p)
}
\arguments{
\item{p}{Numeric vector. Probability vector.}
}
\value{
Numeric value. Normalized Shannon entropy in \eqn{[0, 1]}.
}
\description{
Calculate Normalized Entropy
}
